+++ CHANGELOG.rst
@@ -31,6 +31,7 @@ Added
 Fixed
 =====
 
+- Restore terminal state on unexpected exit signal
 - Prevent ``KeyError: 'log_file'`` error when launching apps via ``euporie`` app
 - Prevent unknown markdown code block language causing rendering error
 - Fix unexpected cropped graphics when using ``timg`` with character aspects ≠ 0.5